refactor(eve): extract continuation bundle overlay
